webhooks: Move github/ to github_legacy/ and remove docs.
