+++ zerver/webhooks/github_legacy/fixtures/v1_issues_opened.json
