+++ zerver/webhooks/github_legacy/fixtures/v2_issues_opened.json
